Improved log method; fixed ReleaseDelayOserver timerId assignment
